+++ Assets/scripts/random_message.js
@@ -24,7 +24,7 @@ for (let index = 0; index < array.length; index++) {
   const element = array[index];
   const randoI = randomInt(element.min, element.max);
   const randoF = random(element.min, element.max);
-  const CAN_CHANGE_MSG = ((element.float) : randoF == element.wanted ? randoI == element.wanted) && new_message == DEFAULT_MESSAGE_INDEX;
+  const CAN_CHANGE_MSG = ((element.float) ? randoF == element.wanted : randoI == element.wanted) && new_message == DEFAULT_MESSAGE_INDEX;
 
   if (CAN_CHANGE_MSG) {
     new_message = element.msg;