+++ Assets/scripts/version-page.js
@@ -1,6 +1,6 @@
 import { returnVersion } from "./utilities/version-utility.js";
 
-var version = returnVersion('version-dev');
 var element = document.getElementById('version');
+var versionString = returnVersion('version-dev');
 
-element.innerHTML = `Dragging Site: ${version}`;
+element.innerHTML = `Dragging Site: ${versionString}`;