+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Github Page
 - JSONs everywhere
   - Random Messages
+  - TODO
 - New Theme
 - Todo list and page